Refactor Raft WAL code into multiple files.
